fix: race condition of pre-closure of progress bar before finish
